requirements.txt
pandas
meteostat
pyspark
python-dotenv